+++ packages/components/src/components/button-group/button-group-styles.spec.ts
@@ -1,6 +1,7 @@
 import { getComponentCss } from './button-group-styles';
 import type { ButtonGroupDirection } from './button-group-utils';
 import type { BreakpointCustomizable } from '../../types';
+import { validateCssAndMatchSnapshot } from '../../../tests/unit/helpers';
 
 describe('getComponentCss()', () => {
   it.each<BreakpointCustomizable<ButtonGroupDirection>>([
@@ -9,6 +10,6 @@ describe('getComponentCss()', () => {
     'row',
     { base: 'row', xs: 'column', s: 'row', m: 'column', l: 'row', xl: 'column' },
   ])('should return correct css for direction: %j', (direction) => {
-    expect(getComponentCss(direction)).toMatchSnapshot();
+    validateCssAndMatchSnapshot(getComponentCss(direction));
   });
 });