+++ packages/components/src/components/flyout-navigation/flyout-navigation-item/flyout-navigation-item-styles.spec.ts
@@ -1,4 +1,5 @@
 import { getComponentCss } from './flyout-navigation-item-styles';
+
 describe('getComponentCss()', () => {
   it.each<Parameters<typeof getComponentCss>>([
     [false, 'light'],
@@ -8,6 +9,7 @@ describe('getComponentCss()', () => {
     [false, 'auto'],
     [true, 'auto'],
   ])('should return correct css for isSecondaryScrollerVisible: %s and theme: %s', (...args) => {
+    // TODO: Use validateCssAndMatchSnapshot when visibility is changed to inherit
     expect(getComponentCss(...args)).toMatchSnapshot();
   });
 });