+++ packages/components/src/components/link-tile/link-tile-styles.spec.ts
@@ -1,4 +1,5 @@
 import { getComponentCss } from './link-tile-styles';
+import { validateCssAndMatchSnapshot } from '../../../tests/unit/helpers';
 
 describe('getComponentCss()', () => {
   it.each<Parameters<typeof getComponentCss>>([
@@ -19,7 +20,7 @@ describe('getComponentCss()', () => {
   ])(
     'should return correct css for aspectRatio: %j, size: %j, weight: %j, background: %s, align: %s, isCompact: %j and hasGradient: %s',
     (...args) => {
-      expect(getComponentCss(...args)).toMatchSnapshot();
+      validateCssAndMatchSnapshot(getComponentCss(...args));
     }
   );
 });