+++ packages/storefront/tests/e2e/specs/navigation.e2e.ts
@@ -8,6 +8,13 @@ import * as fs from 'fs';
 const console = require('console'); // workaround for nicer logs
 let browserPage: Page;
 
+// style overrides for css variables
+const styleOverrides = fs.readFileSync(
+  path.resolve(require.resolve('@porsche-design-system/shared'), '../css/styles.css'),
+  'utf8'
+);
+const [, rootStyles] = /(:root {[\s\S]+?})/.exec(styleOverrides) || [];
+
 beforeEach(async () => {
   browserPage = await browser.newPage();
   initConsoleObserver(browserPage);
@@ -18,122 +25,133 @@ beforeEach(async () => {
 });
 afterEach(async () => await browserPage.close());
 
-const isTabActive = (element: ElementHandle<HTMLElement> | null): Promise<boolean> =>
-  element.evaluate((el) => el.className.includes('router-link-active'));
-const isLinkActive = (element: ElementHandle<HTMLElement> | null): Promise<boolean> =>
-  element.evaluate((el) => el.active);
-const getMainTitle = (page: Page): Promise<string> => page.$eval('.vmark > h1', (x) => x.innerHTML);
-const hasPageObjectObject = (page: Page): Promise<boolean> =>
-  page.evaluate(() => document.body.innerText.includes('[object Object]'));
-
-/**
- * to override transition duration of accordion
- */
 const injectCSSOverrides = async () => {
-  const stylesPath = path.resolve(require.resolve('@porsche-design-system/shared'), '../css/styles.css');
-  const styles = fs.readFileSync(stylesPath, 'utf8');
-
-  await browserPage.evaluate((cssStyles: string) => {
-    const styleTag = document.createElement('style');
-    styleTag.innerText = cssStyles;
-    document.head.appendChild(styleTag);
-  }, styles);
+  // inject style overrides for css variables to override transition duration of accordion
+  await page.evaluate((styles) => {
+    const styleEl = document.createElement('style');
+    styleEl.innerText = styles;
+    document.head.append(styleEl);
+  }, rootStyles);
 };
 
-/*
-// transform STOREFRONT_CONFIG into tuple array with structure [category, page, tab?][]
-const cases: [string, string, string?][] = Object.entries(STOREFRONT_CONFIG)
-  .filter((_, i) => i < 3)
+const isTabActive = (element: ElementHandle<HTMLElement> | null): Promise<boolean> => {
+  return element.evaluate((el) => el.className.includes('router-link-active'));
+};
+
+const isLinkActive = (element: ElementHandle<HTMLElement> | null): Promise<boolean> => {
+  return element.evaluate((el) => el.active);
+};
+
+const waitForHeading = async (page: Page): Promise<ElementHandle<HTMLElement>> => {
+  const headingElementHandle = (await page.waitForSelector('main .vmark > h1')) as ElementHandle<HTMLElement>;
+  // NOTE: sometimes h1 or p-heading is rendered empty for whatever reason 🤷‍
+  await page.waitForFunction((headingEl) => headingEl.innerText !== '', undefined, headingElementHandle);
+  return headingElementHandle;
+};
+
+const getHeadingText = async (page: Page): Promise<string> => {
+  const headingElementHandle = await waitForHeading(page);
+  return headingElementHandle.evaluate((headingEl) => headingEl.innerText);
+};
+
+const hasPageObjectObject = (page: Page): Promise<boolean> => {
+  return page.evaluate(() => document.body.innerText.includes('[object Object]'));
+};
+
+// transform STOREFRONT_CONFIG into tuple array with structure [category, page, tab, isFirst][]
+const cases: [string, string, string | undefined, boolean][] = Object.entries(STOREFRONT_CONFIG)
+  // .filter((_, i) => i < 2) // NOTE: for easier debugging and testing
   .map(([category, pages]) =>
     Object.entries(pages)
-      .map<[string, string, string?][]>(([page, tabs]) =>
+      .map<[string, string, string | undefined, boolean][]>(([page, tabs]) =>
         typeof tabs === 'object' && !Array.isArray(tabs)
-          ? Object.keys(tabs).map((tab) => [category, page, tab])
-          : [[category, page, undefined]]
+          ? Object.keys(tabs).map((tab, i) => [category, page, tab, i === 0])
+          : [[category, page, undefined, false]]
       )
       .flat()
   )
   .flat();
 
-it.each(cases.map((segments) => [segments.filter(Boolean).join(' > '), ...segments]))(
-  'should navigate to "%s" and have correct heading',
-  (_, category, page, tab) => {
-    // WIP: first part with page is redundant when tabs do exist, maybe route directly to tab without clicking through tabs bar?
-    console.log(category, page, tab);
+it.each<[string, string, string, string | undefined, boolean]>(
+  cases.map((segments) => [segments.filter((segment) => typeof segment === 'string').join(' > '), ...segments])
+  // .filter((_, i) => i < 20) // NOTE: for easier debugging and testing
+)('should navigate to "%s" and have correct heading', async (_, category, page, tab, isFirst) => {
+  const caseIndex = cases.findIndex(
+    ([itemCategory, itemPage, itemTab]) => itemCategory === category && itemPage === page && itemTab === tab
+  );
+  const humanCaseIndex = caseIndex + 1;
+  const counter = `${Array.from(Array(cases.length.toString().length - humanCaseIndex.toString().length))
+    .map(() => ' ') // add leading spaces if needed for nice formatting
+    .join('')}${humanCaseIndex}/${cases.length}`;
+  console.log(`${counter}: ${[category, page, tab].filter(Boolean).join(' > ')}`);
+
+  const [accordionElement] = (await browserPage.$x(
+    `//div[contains(@class, 'menu-desktop')]//nav/p-accordion[@heading='${category}']`
+  )) as ElementHandle<HTMLElement>[];
+  await accordionElement.click();
+  await browserPage.waitForFunction(
+    (el) => getComputedStyle(el.shadowRoot.querySelector('.collapsible')).visibility === 'visible',
+    undefined,
+    accordionElement
+  );
+
+  // reconstruct href of p-link-pure in sidebar
+  // for first tab it is correct
+  // for other tabs we need to use first tab
+  // for everything else there is no tab url
+  const firstOrNoTab =
+    tab &&
+    (isFirst
+      ? tab
+      : cases.find(([itemCategory, itemPage, itemTab]) => itemCategory === category && itemPage === page)[2]);
+  const linkPureHref = `/${[category, page, firstOrNoTab]
+    .filter(Boolean)
+    .map((part) => paramCase(part))
+    .join('/')}`;
+  const [linkPureElement] = (await browserPage.$x(
+    `//div[contains(@class, 'menu-desktop')]//nav//p-link-pure/a[contains(., '${page}')][@href='${linkPureHref}']/parent::p-link-pure`
+  )) as ElementHandle<HTMLElement>[];
+  expect(await isLinkActive(linkPureElement), 'sidebar link should not be active initially').toBe(false);
+
+  // NOTE: very flaky and potential timeout here 🤷‍
+  await Promise.all([browserPage.waitForNavigation(), linkPureElement.click()]);
+
+  // wait for p-heading and p-tabs-bar to be ready
+  const mainElementHandle = await browserPage.$('main');
+  await mainElementHandle.evaluate((el) => (window as any).componentsReady(el));
+
+  await waitForHeading(browserPage);
+  await browserPage.waitForFunction((el) => el.active, undefined, linkPureElement);
+  expect(await isLinkActive(linkPureElement), 'sidebar link should be active after click').toBe(true);
+
+  const headingRegEx = new RegExp(`^${page}( 🚫| 🧪)?$`); // to cover deprecated and experimental icon
+  expect(await getHeadingText(browserPage), 'should show correct main title for page').toMatch(headingRegEx);
+  expect(await hasPageObjectObject(browserPage), 'should not contain [object Object] on page').toBe(false);
+  expect(getConsoleErrorsAmount(), `Errors on ${category}/${page}`).toBe(0);
+
+  if (tab) {
+    const tabHref = `\/${paramCase(category)}\/${paramCase(page)}\/${paramCase(tab)}`;
+    const [tabElement] = (await browserPage.$x(
+      `//p-tabs-bar//a[contains(., '${tab}')][@href='${tabHref}']`
+    )) as ElementHandle<HTMLElement>[];
+
+    const isTabElementActiveInitially = await isTabActive(tabElement);
+    if (isFirst) {
+      expect(isTabElementActiveInitially, 'should have first tab active initially').toBe(true);
+      // heading of first is already checked before
+    } else {
+      expect(isTabElementActiveInitially, 'should not have tab active initially').toBe(false);
+      // we need to switch tabs, e.g. to "Usage" or "Props" for components
+
+      await Promise.all([browserPage.waitForNavigation(), tabElement.click()]);
+
+      expect(await isTabActive(tabElement), 'should have tab active after click').toBe(true);
+      expect(await isLinkActive(linkPureElement), 'sidebar link should still be active after click').toBe(true);
+
+      await waitForHeading(browserPage);
+      expect(await getHeadingText(browserPage), 'should show correct main title for tab page').toMatch(headingRegEx);
+      expect(await hasPageObjectObject(browserPage), 'should not contain [object Object] on tab page').toBe(false);
+      expect(getConsoleErrorsAmount(), `Errors on ${category}/${page} in tab ${tab}`).toBe(0);
+    }
   }
-);
-*/
-
-// TODO: we should prepare a test array and loop by it.each instead to have better console output while test case is executed
-// see above for a WIP
-for (const [category, pages] of Object.entries(STOREFRONT_CONFIG)) {
-  for (const [page, tabs] of Object.entries(pages)) {
-    ((category: string, page: string) => {
-      const tab = !Array.isArray(tabs) ? Object.keys(tabs)[0] : undefined;
-      it(`should navigate to "${category} > ${page} ${tab !== undefined ? '> ' + tab : ''}"`, async () => {
-        console.log(`${category} > ${page} ${tab !== undefined ? '> ' + tab : ''}`);
-        const [accordionElement] = (await browserPage.$x(
-          `//div[contains(@class, 'menu-desktop')]//nav/p-accordion[@heading='${category}']`
-        )) as ElementHandle<HTMLElement>[];
-        await accordionElement.click();
-
-        const href = `\/${paramCase(category)}\/${paramCase(page)}${tab !== undefined ? '/' + paramCase(tab) : ''}`;
-        const [linkPureElement] = (await browserPage.$x(
-          `//div[contains(@class, 'menu-desktop')]//nav//p-link-pure/a[contains(., '${page}')][@href='${href}']/parent::p-link-pure`
-        )) as ElementHandle<HTMLElement>[];
-        expect(await isLinkActive(linkPureElement), 'sidebar link should not be active initially').toBe(false);
-
-        await Promise.all([browserPage.waitForNavigation(), linkPureElement.click()]);
-
-        const mainElementHandle = await browserPage.$('main');
-        await mainElementHandle.waitForSelector('.vmark > h1');
-        expect(await isLinkActive(linkPureElement), 'sidebar link should be active after click').toBe(true);
-        expect(await getMainTitle(browserPage), 'should show correct main title for page').toBe(page);
-        expect(await hasPageObjectObject(browserPage), 'should not contain [object Object] on page').toBe(false);
-        expect(getConsoleErrorsAmount(), `Errors on ${category}/${page}`).toBe(0);
-
-        if (!Array.isArray(tabs)) {
-          // wait for p-tabs-bar to be ready
-          await mainElementHandle.evaluate((el) => (window as any).componentsReady(el));
-
-          for (const [index, tab] of Object.entries(Object.keys(tabs))) {
-            console.log(`– ${tab}`);
-            const tabHref = `\/${paramCase(category)}\/${paramCase(page)}\/${paramCase(tab)}`;
-            const [tabElement] = (await browserPage.$x(
-              `//p-tabs-bar//a[contains(., '${tab}')][@href='${tabHref}']`
-            )) as ElementHandle<HTMLElement>[];
-            const isFirstTab = parseInt(index) === 0;
-
-            const isTabElementActiveInitially = await isTabActive(tabElement);
-            if (isFirstTab) {
-              expect(isTabElementActiveInitially, 'should have first tab active initially').toBe(true);
-            } else {
-              expect(isTabElementActiveInitially, 'should not have tab active initially').toBe(false);
-
-              // const articleHTML = await mainElementHandle.evaluate(
-              //   (mainEl) => mainEl.querySelector('article').innerHTML
-              // );
-
-              await Promise.all([browserPage.waitForNavigation(), tabElement.click()]);
-
-              // await browserPage.waitForFunction(
-              //   (mainEl, html) => mainEl.querySelector('article').innerHTML !== html,
-              //   {},
-              //   mainElementHandle,
-              //   articleHTML
-              // );
-              expect(await isTabActive(tabElement), 'should have tab active after click').toBe(true);
-            }
-
-            await mainElementHandle.waitForSelector('.vmark > h1');
-            expect(await getMainTitle(browserPage), 'should show correct main title for tab page').toBe(page);
-            expect(await hasPageObjectObject(browserPage), 'should not contain [object Object] on tab page').toBe(
-              false
-            );
-            expect(getConsoleErrorsAmount(), `Errors on ${category}/${page} in tab ${tab}`).toBe(0);
-          }
-        }
-      });
-    })(category, page);
-  }
-}
+});